Fix warnings

+++ Team18/Code18/source/QPS/Evaluators/ConnectedSynonymClauseGroup.cpp
@@ -25,7 +25,7 @@ bool spa::operator==(const ConnectedSynonymClauseGroup& first, const ConnectedSy
     return false;
   }
 
-  for (int i = 0; i < first.clauses.size(); ++i) {
+  for (size_t i = 0; i < first.clauses.size(); ++i) {
     if (first.clauses[i] != second.clauses[i]) {
       return false;
     }
